pyproject.toml
[project]
name = "vlm-web-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "python-dotenv>=1.0.1",
]
